pyproject.toml
[project]
name = "book-scraper"
version = "0.1.0"
description = "Async book crawler for books.toscrape.com"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Contributors" }]
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "pymongo>=4.6.0",
  "beautifulsoup4>=4.12.0",
  "apscheduler>=3.10.4",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-multipart>=0.0.6",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "slowapi>=0.1.9"
]

[tool.pdm]

[tool.pdm.dev-dependencies]
test = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.7",
  "coverage>=7.6.0",
  "respx>=0.21.1"
]
lint = [
  "ruff>=0.6.0",
  "black>=24.8.0",
  "mypy>=1.10.0"
]

[tool.pdm.scripts]
crawl = { cmd = "python -m src.crawler.cli crawl --resume" }
scheduler = { cmd = "python -m src.scheduler.cli run" }
detect = { cmd = "python -m src.scheduler.cli detect" }
report = { cmd = "python -m src.scheduler.cli report" }
api = { cmd = "uvicorn src.api.app:app --reload --host 0.0.0.0 --port 8000" }
test = { cmd = "pytest -q --maxfail=1 --disable-warnings" }
lint = { cmd = "ruff check ." }
format = { cmd = "black ." }
typecheck = { cmd = "mypy src" }

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

